# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "測試 Swagger API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:7001/",
            "description": "localhost"
        }
    ],
    "paths": {
        "/api/city": {
            "get": {
                "tags": [
                    "city Tag"
                ],
                "summary": "取得 city",
                "operationId": "city",
                "responses": {
                    "200": {
                        "description": "請求成功"
                    },
                    "404": {
                        "description": "請求失敗"
                    }
                }
            },
            "post": {
                "tags": [
                    "city Tag"
                ],
                "summary": "新增 city",
                "operationId": "create city",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "名稱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    },
                    "422": {
                        "description": "格式內容錯誤"
                    },
                    "404": {
                        "description": "請求失敗"
                    }
                }
            }
        },
        "/api/city/{id}": {
            "get": {
                "tags": [
                    "city Tag"
                ],
                "summary": "取得 city 編號",
                "operationId": "city id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "內容",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    },
                    "404": {
                        "description": "請求失敗"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "city Tag"
                ],
                "summary": "編輯 city",
                "operationId": "update city",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "編號",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "名稱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    },
                    "422": {
                        "description": "格式內容錯誤"
                    },
                    "404": {
                        "description": "請求失敗"
                    }
                }
            },
            "delete": {
                "tags": [
                    "city Tag"
                ],
                "summary": "刪除 city",
                "operationId": "delete city",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "編號",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "請求成功"
                    },
                    "404": {
                        "description": "請求失敗"
                    }
                }
            }
        },
        "/api/create": {
            "get": {
                "operationId": "c9a1b4317fbd14ad85095de576004941",
                "responses": {
                    "200": {
                        "description": "create text."
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}